refactor: centralize llm helpers
