+++ movie_agent/image_ui.py
@@ -19,12 +19,8 @@ from movie_agent.comfyui import (
     DEFAULT_STEPS,
     DEFAULT_NEGATIVE_PROMPT,
 )
-from movie_agent.ollama import (
-    list_ollama_models,
-    generate_story_prompt,
-    DEBUG_MODE,
-)
-from movie_agent.lmstudio import generate_story_prompt_lmstudio, list_lmstudio_models
+from movie_agent.ollama import DEBUG_MODE
+from movie_agent.llm_helpers import select_llm_models, generate_prompt_for_row
 from movie_agent.csv_manager import (
     load_image_data,
     save_data,
@@ -175,47 +171,6 @@ def fetch_view_counts(site: str, post_id: str) -> Dict[str, int]:
         views = data.get("views", [])
         results[column] = views[0] if views else 0
     return results
-
-
-def select_llm_models(df: pd.DataFrame) -> list[str]:
-    """Return available LLM models based on ``llm_environment`` column."""
-    if (
-        "llm_environment" in df.columns
-        and df["llm_environment"].astype(str).str.lower().eq("lmstudio").any()
-    ):
-        return list_lmstudio_models()
-    return list_ollama_models()
-
-
-def generate_prompt_for_row(
-    row: pd.Series,
-    context: str,
-    model: str,
-    temperature: float,
-    max_tokens: Optional[int],
-    top_p: Optional[float],
-    timeout: int,
-) -> Optional[str]:
-    """Generate a prompt using the appropriate LLM backend."""
-    env = str(row.get("llm_environment", "")).strip().lower()
-    if env == "lmstudio":
-        return generate_story_prompt_lmstudio(
-            context,
-            model,
-            temperature=temperature,
-            max_tokens=max_tokens,
-            top_p=top_p,
-            timeout=timeout,
-        )
-    return generate_story_prompt(
-        context,
-        model=model,
-        temperature=temperature,
-        max_tokens=max_tokens,
-        top_p=top_p,
-        debug=DEBUG_MODE,
-        timeout=timeout,
-    )
 
 
 def main() -> None:
